+++ src/camera.h
@@ -1,5 +1,7 @@
 #pragma once
 #include <raylib.h>
+#include <cstdint>
+#include <sys/types.h>
 
 #include "input.h"
 
@@ -10,17 +12,13 @@ struct CenteredCamera {
     float rotation = 0.;
     float zoom = 2.;
 
-    Camera2D GetCamera2D(unsigned int window_width, unsigned int window_height) const;
+    Camera2D GetCamera2D(uint16_t window_width, uint16_t window_height) const;
 
-    Rectangle GetBounds(unsigned int window_width, unsigned int window_height) const;
+    Rectangle GetBounds(uint16_t window_width, uint16_t window_height) const;
 
-    void UpdateZoom(float mouse_wheel_input, bool ctrl_held, unsigned int window_width, unsigned int window_height);
+    void UpdateZoom(float mouse_wheel_input, bool ctrl_held, uint16_t window_width, uint16_t window_height);
 
     void UpdatePosition(Vector2 player_center, float window_width, float window_height);
 
-    void Update(Vector2 player_center,
-        Input input,
-        float window_width,
-        float window_height
-    );
+    void Update(Vector2 player_center, Input input, float window_width, float window_height);
 };
